Python: add templates to recognize keywords in sections
Each template file is a dictionary with keys for keywords for that
section in the input file. Each keyword can produce multiple outputs,
using utility methods to process the parse results nested list.

Still Todo:
Little or no verification is done, and all cells and lattices are
output, whether they are used or not.
Several sections not related to reactor geometry are not done.