Add conditional output of cells and cellmaps.
Use a 'reference parameter' which is saved when a card is output
to check whether to output a cell or cellmap based on their use
in axial or cellmap cards.

Combine common parts of sections that have cells and cellmaps into
a single file, and just specify the unique parts per section.