Add verification, refactor cell-maps, add more blocks
Add a basic verification step, so when a card is encountered
it can run verification on its contents. Only added to a few
cards so far.

Factor out the common parts of blocks that have cell-maps, and
re-use it by merging dictionaries

Add keyword lists for simulator-specific blocks. Most are defaulted
to just copy a single value, so output is wrong in many cases so far.
Some handling of a simple list, where a few cards are grouped in
a list in the xml output.