Template conversion for Simput
Add a script that takes the python templates for the simulation
config blocks, and outputs a json object representing the template.
It replaces functions by their name, so it is serializable.

Simput will use the json objects to automatically construct a UI for
setting all the parameters.

Fix a confusing variable name in Utils.py